fixed many bugs from this weekends updating but there are still a handful of bugs to go through but good progress recently

+++ src/Imaginengion/Renderer/RenderPipeline.zig
@@ -29,8 +29,8 @@ pub const SDFPushConstants = extern struct {
     mRayOffset: Vec2(f32).VectorT,
     mQuadsCount: u32,
     mGlyphsCount: u32,
-    mViewportWidth: u32,
-    mViewportHeight: u32,
+    mViewportWidth: f32,
+    mViewportHeight: f32,
 };
 
 pub fn Pipeline(pipeline_t: PipelineType) type {
